# src/main/java/com/fulfilment/application/monolith/warehouses/domain/models/Warehouse.java
package com.fulfilment.application.monolith.warehouses.domain.models;

import java.time.LocalDateTime;

public class Warehouse {

  // database id (used by GET/DELETE /warehouse/{id})
  public Long id;

  // unique business code for the active warehouse
  public String businessUnitCode;

  public String location;

  public Integer capacity;

  public Integer stock;

  public LocalDateTime createdAt;

  public LocalDateTime archivedAt;
}